this checkin adds the main block of code for the new event system, fixes for the bootstrap required to get it working, and a smoke-test for the new event system as well. Since the full testharness isn't yet working from ant, the test file uses hard-coded load() calls to include the correct files. This will need to be re-visited.
Also, the location of classes and items in the dojo.event namespace need to be re-checked and aliases and wrappers to dojo.MOP should be created since we are essentially using MOP to do the work of eventing in this case.

The browser side of the event system will be comitted separately.


git-svn-id: http://svn.dojotoolkit.org/src@78 560b804f-0ae3-0310-86f3-f6aa0a117693