Don't do dojo.require("dojo.widget.DatePicker.util"); It tries to pull in the file called DatePicker/util.js, which doesn't exist.
Although this works in a filesystem environment but when accessing over a webserver (at
least dojotoolkit.org), the server returns an HTML page saying "file not found" in lieu of
a javascript file, which naturally causes the browser to print some error message about
invalid javascript.

Try demos/Mail.html for a testcase.

git-svn-id: http://svn.dojotoolkit.org/src/trunk@2796 560b804f-0ae3-0310-86f3-f6aa0a117693