Fix bug in splitpane's css (I think).  Fixes #495.
SplitPane supports a margin around each of it's children, but it
was specifying the margin by setting "padding" rather than "margin",
which was messing up TabPane.

If anything, .dojoHtmlSplitterPanePanel should specify a margin.
Specifying padding doesn't make sense.  But I decided not to specify
anything, and instead just let the user specify.

git-svn-id: http://svn.dojotoolkit.org/src@3249 560b804f-0ae3-0310-86f3-f6aa0a117693