Fixes for floating pane on IE, so the resize bar doesn't appear misaligned.
IE runs in strict mode or compatibility mode (compatibility mode is the default),
depending on the header at the top of the html page.
One difference is whether or not width:100% includes the borders.

Ideally dojo should support both modes, but I think to do that we will need
to use css expressions to do that (?).  Anyway, for now I am (reluctantly)
checking in a version that supports IE in compatibility mode only.

Subject to change later.

git-svn-id: http://svn.dojotoolkit.org/src/trunk@3258 560b804f-0ae3-0310-86f3-f6aa0a117693