Lots of fixes for ContentPane remote loading, courtesy of Fredrik Johansson (CLA on file).
In particular this patch adjusts paths for <img src="...">, <a href="...">, etc. to be
relative to the containing page, so that image and links work properly.

This patch should allow people to use
  <div dojoType="ContentPane" href="..."></div>
where they previously had to use an <iframe>.  (Maybe people won't try to use
iframes inside SplitPanes anymore?)  Nice work!

git-svn-id: http://svn.dojotoolkit.org/src@3289 560b804f-0ae3-0310-86f3-f6aa0a117693