Rev# 3365 (Ticket#534) suddenly changed the signature for dojo.lang.forEach() so the third argument became a number.
In the original, the third
argument is a boolean which if true specifies that the
callback function guarantees not to change the length
of the array.

I'm not sure that boolean flag was useful in the first
place (it's a dubious performance improvement), but event.js and
other code pass "true" as the third argument, so you can't
just change the signature.

Reverting to the old function signature.


git-svn-id: http://svn.dojotoolkit.org/src/trunk@3372 560b804f-0ae3-0310-86f3-f6aa0a117693