Rework the resizing notification code (onResized()) to avoid hangs, and hopefully improve performance, by only doing resizing for visible widgets, and only if the widget's size has actually changed.
- If a parent widget changes it's size, it can either size
it's child widgets manually by calling resizeTo(w,h), or
just tell them that their size has possibly changed by calling
onParentResized().

onParentResized() is for the benefit of child widgets that are
relatively sized (like height: 100%), and that need to know
when their parent's size changes.

onParentResized() checks to make sure that the widget is
visible and the size has really changed, and if so calls
onResized().

- lazy sizing: if you resize the browser window, TabContainer
will only resize the currently displayed tab.  As a corollary,
when you call show() on a widget, the widget should check to
see if it's size has changed.

- bgIframe is resized in the individual widget code, not in
HtmlWidget

git-svn-id: http://svn.dojotoolkit.org/src@3645 560b804f-0ae3-0310-86f3-f6aa0a117693