Hopefully fix button sizing issues.
You can't compute the size of a node if any of it's ancestors
are hidden (display:none), so I need to attach to document.body
whenver computing size.  This is true not only for initialization,
but also when the user changes the button caption.

Should probably put this code in a more general place, but
putting it in Button for now.


git-svn-id: http://svn.dojotoolkit.org/src/trunk@4186 560b804f-0ae3-0310-86f3-f6aa0a117693