Temporary fix for #870: Combobox dropdown occasionally does not work
The checkBlurred code is runs on a timer and actually was firing while the menu was fading in (or on a timer to fade in?).  I changed the timer values so it wouldn't fire until after the drop down was fully displayed.  This code is still a bit of a mystery to me but this change seems to make the problem go away.


git-svn-id: http://svn.dojotoolkit.org/src/trunk@4287 560b804f-0ae3-0310-86f3-f6aa0a117693