fixes #1525: Subwidget definitions not necessarily respecting CSS definitions.
The problem is not failure of respecting CSS. The real issue is that when the chidren widget are created, the top widget is not appended to the document dom tree, so its size is invalid and LayoutContainer can not layout its children successfully. This patch moves the widgetsInTemplate support after this.domNode is appended to the dom tree.

In order to have this work correctly, dojo.widget.createWidget should be called with a third parameter to specify where to insert this widget in the dom tree. This is required if you use some widgets (like LayoutContainer) relying on correct sizing when created.

git-svn-id: http://svn.dojotoolkit.org/src@5911 560b804f-0ae3-0310-86f3-f6aa0a117693