This changes the treatment of the virtual sizer from hide/show to move to back/bring to front. We play with the zIndex instead of display property.
This is a variant of the 'quirky percentages in IE' issue (see http://www.positioniseverything.net/explorer/percentages.html ) that comes about when you use relative positioning and percentages for dimensions. Toggling display and/or visibility induces the behaviour in Internet Explorer.

Patch from Koranteng Ofosu-Amaah (IBM, CCLA on file)

git-svn-id: http://svn.dojotoolkit.org/src/trunk@7313 560b804f-0ae3-0310-86f3-f6aa0a117693