Prototype for simple FocusManager class, to save/restore focus and selection.
Example usage: save the state before displaying a dialog, and then restore it
after the dialog is dismissed.

Class tries to keep track of the previously
focused element in addition to the current one because (for example) clicking
a toolbar will cause the toolbar to receive focus before this class is called,
but we need to get back to the state before the toolbar was focused.


git-svn-id: http://svn.dojotoolkit.org/src@7486 560b804f-0ae3-0310-86f3-f6aa0a117693