Intermediate checkin of Tree code (based on Dojo 0.4 TreeV3, but without DnD, editing, or selection).
The code is split into Tree, the visual part, and Controller.
Controller responds to events from the Tree.   For instance,
when the user  clicks the +/- expando icon, the tree sends a
"toggleOpen" event, and then the Controllerqueries the dojo.data.Store
to get the data and sends it to tree to display it.

Nothing is working yet; just an intermediate checkin.

Refs #2782

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@8278 560b804f-0ae3-0310-86f3-f6aa0a117693