Fixes so tooltip runs on IE.
IE overflow:visible doesn't work on a node w/an opacity setting, so make the node fully visible.

Also, the filter to display a png image is problematic because the path has to be relative
to the main document, not to the CSS file, and we don't know where that is.   Using a gif file
on IE6 for now, as a compromise.

Also, specify the opacity in CSS rather than as a widget parameter.

Refs #2639
 

git-svn-id: http://svn.dojotoolkit.org/src@8857 560b804f-0ae3-0310-86f3-f6aa0a117693