Fixes for TitlePane url support.
The code in ContentPane to connect to StackContainer's selectChild topic was too specialized.  It didn't work for TitlePane.  (For TitlePane we also need to defer the href load until the pane is opened.)  Changed StackContainer to call Widget.onShow() (if such a method exists) when a pane is selected, and TitlePane to do the same when it's content is opened.  Made TitlePane extend ContentPane.

As a side effect this means that <div dojoType=dijit.layout.ContentPane href=...></div> *not* inside of a StackContainer won't load the data at all... you have to set preload=true.  But that seems like an uncommon case, as it would have been more efficient to just inline the data and avoid the immediate extra trip to the server.

The href for AccordionPane won't work correctly until AccordionPane extends ContentPane (refs #3529)

Fixes #3645 (but I still need to add code to make the animation smoother when the
href load is delayed).

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@9849 560b804f-0ae3-0310-86f3-f6aa0a117693