Better testcase for how I want InlineEditBox to work.
 - shows how you should specify a width for inline elements similar to the way
you specify the width of an <input>... ie, a width that is unrelated to the
width of the current value.

 - Also gives much better test of inherited styles from <style> tags and from the
browser's builtin rules for <p>, <h3>, etc.  (All such style info must appear in
the non-editing version of the text and the edit version of the tet.)

 - Also demonstrates a problem with specifying paragraphs of text, inserting
unwanted line breaks (see #4098 for possible solution).

Refs #4097, #4098, #4167.

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@10354 560b804f-0ae3-0310-86f3-f6aa0a117693