Make tabbing into/out of a MenuBar work.
Had to rearchitect KeyNavContainer to be aware of Containers that spawn popups.  Made the setting/clearing of tabindex on the Container's domNode contigent on the container being in/out of the focus chain.  In other words, focusing on a MenuItem in a submenu will no longer cause the main Container to restore it's tabIndex.

Refs #4680 !strict.

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@16142 560b804f-0ae3-0310-86f3-f6aa0a117693