The new query engine (aka: "steak").
Making claims about the speed of the system would be counter-productive, but
suffice to say, steak is really, really fast.. It acheives its speed in ways
that differ from its predacessor. Instead of branching for DOM vs. XPath, the
new engine branches QSA vs. DOM, dropping the XPath engine entirely. This
allows for a sizable reduction in overall size of the codebase. Next, the query
dispatch function has been moved from a recursive system which always had some
obvious warts to a set of nested loops which have differnt (but easier to
understand and hack on) warts.

Sizable room for improvement performance exists as it may be possible to use
the tokenizer to say more about a particular query and how it should be
dispatched. Also, it may be possible to run more queries through QSA than we
currently do. This is being investigated for 1.3.x.

Lastly, steak is stand-alone. Recent versions of dojo.query have been able to
stand on their own with minimal scaffolding, but steak improves upon that by
building default scaffolding in but hiding it behind build-time ifdefs to
ensure that the size of Dojo isn't affected.

Fixes #7072. Fixes #4425. !strict

Tests coming in next commit.


git-svn-id: http://svn.dojotoolkit.org/src/dojo/trunk@16217 560b804f-0ae3-0310-86f3-f6aa0a117693