When specifying a deferred-connect handler (like onmousemove) as an initialization parameter, and then later doing a dojo.connect(), the connect wasn't working.
Problem was that the connection from the Widget.focusNode to Widget.onMouseMove() was
incorrect; it connected to the actual function, rather than using the string "onMouseMove", so
that the later dojo.connect() call, which replaces Widget.onMouseMove() with a dispatcher
function, wasn't affecting what the Widget.focusNode connection was executing.

Fixes #8814 !strict.

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@16926 560b804f-0ae3-0310-86f3-f6aa0a117693