Make nested ContentPanes defer loading of href and resizing of child widgets until the ContentPane is shown.   Specifically addressing complicated cases like nested TabContainers or TabContainer --> BorderContainer --> ContentPane.
When ContentPane is a child of a layout container it's hard to efficiently determine whether it's shown or not, but we know that resize() will be called the first time it's shown (perhaps on page load), so if resize() _hasn't_ been called then defer loading of href and resizing of child widgets.

Fixes #5528, #9794 !strict. 


git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@20000 560b804f-0ae3-0310-86f3-f6aa0a117693