!ComboButton: lucid needs hover/active CSS change on <td> containing the <button>, rather than just the dijitButtonContentsHover class on the <button> itself.  Thanks Wei, refs #10527 !strict.
It's a little confusing because both the left and right part of a !ComboButton have the dijitButtonNode class (since they are both buttons), but only the left part gets dijitButtonHover, whereas the right part gets dijitDownArrowButtonHover.  Maybe they should both have distinct names, or both have the same name, but I'll leave it this way for now.

git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@21184 560b804f-0ae3-0310-86f3-f6aa0a117693